+++ src/main/kotlin/com/jetbrains/snakecharm/codeInsight/completion/wrapper/SnakemakeStartupActivity.kt
@@ -4,6 +4,7 @@ import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.StartupActivity
 import com.jetbrains.snakecharm.codeInsight.ImplicitPySymbolsProvider
+import com.jetbrains.snakecharm.codeInsight.completion.yamlKeys.SmkYAMLKeysStorage
 import com.jetbrains.snakecharm.framework.SmkSupportProjectSettings
 import kotlinx.serialization.ExperimentalSerializationApi
 
@@ -14,6 +15,8 @@ class SnakemakeStartupActivity : StartupActivity.Background {
         smkSettings.initOnStartup()
         val smkWrapperStorage = project.service<SmkWrapperStorage>()
         smkWrapperStorage.initOnStartup()
+        val smkYamlKeysStorage = project.service<SmkYAMLKeysStorage>()
+        smkYamlKeysStorage.initOnStartup()
 
         val implicitPySymbolsProvider = project.service<ImplicitPySymbolsProvider>()
         implicitPySymbolsProvider.initOnStartup()