+++ src/main/kotlin/com/jetbrains/snakecharm/stringLanguage/lang/psi/SmkSLSubscriptionIndexKeyExpression.kt
@@ -2,4 +2,11 @@ package com.jetbrains.snakecharm.stringLanguage.lang.psi
 
 import com.jetbrains.python.psi.PyReferenceExpression
 
-interface SmkSLSubscriptionIndexKeyExpression : SmkSLReferenceExpression, PyReferenceExpression
+interface SmkSLSubscriptionIndexKeyExpression : SmkSLReferenceExpression, PyReferenceExpression {
+    /**
+     * Returns [true] if index key is empty
+     */
+    // E.g. for 'config[]' -> returns true
+    // 'config[key]' -> returns false
+    fun hasEmptyIndex(): Boolean
+}